refactor(frontend): consolidate getCssVar + add getSeverityColor to shared composable (#1606, #1647)
Migrate 18 local getCssVar() copies to shared useCssVars.ts composable.
Add getSeverityColor() and getSeverityColors() as shared utilities
for theme-aware severity color resolution across charts/dashboards.
Make fallback parameter optional (default '') for callers that omit it.
