+++ autobot-frontend/src/components/charts/ModuleImportsChart.vue
@@ -23,17 +23,9 @@ import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 import BaseChart from './BaseChart.vue'
 import type { ApexOptions } from 'apexcharts'
+import { getCssVar } from '@/composables/useCssVars'
 
 const { t } = useI18n()
-
-/**
- * Get CSS variable value from the document
- * Issue #704: Use design tokens for theming
- */
-function getCssVar(name: string, fallback: string): string {
-  if (typeof document === 'undefined') return fallback
-  return getComputedStyle(document.documentElement).getPropertyValue(name).trim() || fallback
-}
 
 interface ModuleData {
   path: string
